+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## Unreleased
+
+* Added: FortiNet client VPN cmdlets
+* Added: Windows VPN cmdlets
+
+
 ## 2.4.0 - 2018-10-01
 
 * Added: Add Use-VaultSecureString function